+++ deployer/6/set.php
@@ -74,15 +74,15 @@ set('db_databases',
         'database_default' => [
             get('db_default'),
             !empty($_ENV['IS_DDEV_PROJECT']) ? get('db_ddev_database_config') :
-            (new \SourceBroker\DeployerExtendedTypo3\Drivers\Typo3EnvDriver)->getDatabaseConfig(
-                [
-                    'host' => 'TYPO3__DB__host',
-                    'port' => 'TYPO3__DB__port',
-                    'dbname' => 'TYPO3__DB__database',
-                    'user' => 'TYPO3__DB__username',
-                    'password' => 'TYPO3__DB__password',
-                ]
-            ),
+                (new \SourceBroker\DeployerExtendedTypo3\Drivers\Typo3EnvDriver)->getDatabaseConfig(
+                    [
+                        'host' => 'TYPO3__DB__host',
+                        'port' => 'TYPO3__DB__port',
+                        'dbname' => 'TYPO3__DB__database',
+                        'user' => 'TYPO3__DB__username',
+                        'password' => 'TYPO3__DB__password',
+                    ]
+                ),
         ]
     ]
 );