+++ deployer/default/deploy/task/deploy.php
@@ -20,7 +20,10 @@ task('deploy', [
     'deploy:check_branch',
 
     // Standard deployer task.
-    'deploy:prepare',
+    'deploy:check_remote',
+
+    // Standard deployer task.
+    'deploy:setup',
 
     // Standard deployer task.
     'deploy:lock',
@@ -77,13 +80,10 @@ task('deploy', [
     'deploy:unlock',
 
     // Standard deployer task.
-    'cleanup',
-
-    // Read more on https://github.com/sourcebroker/deployer-extended#deploy-extend-log
-    'deploy:extend_log',
+    'deploy:cleanup',
 
     // Standard deployer task.
-    'success',
+    'deploy:success',
 
 ])->desc('Deploy your TYPO3');