+++ deployer/default/set.php
@@ -33,7 +33,6 @@ set('clear_paths', [
     '.php_cs',
     '.php-cs-fixer.php',
     'dynamicReturnTypeMeta.json',
-    'composer.json',
     'composer.lock',
     'composer.phar',
     'phive.xml',